\name{Birth}
\alias{Birth}
\docType{data}
\title{
Birth Weight data
}
\description{
Classical Birth Weight data }
\usage{data("Birth")}
\format{
  A data frame with 189 observations on the following 6 variables.
  \describe{
    \item{\code{Birth_wt}}{a numeric vector}
    \item{\code{Mom_wt}}{a numeric vector}
    \item{\code{Mom_ht}}{a numeric vector}
    \item{\code{Mom_Age}}{a numeric vector}
    \item{\code{Religion}}{a numeric vector, 1:Hindu 2:Muslim 3:Christian}
    \item{\code{Gender}}{a numeric vector 1 indicates Male and 2 Female}
  }
}
\details{
Can be used for learning and exploring linear regression
}
\source{
Data is a Classical data used to teach linear regression in Manipal Biostatistics Department
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(Birth)
## maybe str(Birth) ; plot(Birth) ...
}
\keyword{datasets}
