% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching.R
\name{branching}
\alias{branching}
\title{branching process}
\usage{
branching(X_o, p, n = NULL)
}
\arguments{
\item{X_o}{the initial population size}

\item{p}{probabilty parameter}

\item{n}{number of generations (Dont Specify if you want the number of generation before extintion  )}
}
\description{
branching process
}
\author{
Nithin Nayak
}
