% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounderPlot.R
\name{confounderPlot}
\alias{confounderPlot}
\title{foo}
\usage{
confounderPlot(data, S, x, y, labels, method = c("associationMeasures",
  "distcor"), returnS = FALSE, plotLegend = TRUE, col, pch, font,
  cex.text, xlim, ylim, ...)
}
\description{
foo
}
