% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CowBullgame.R
\name{CowBull.game}
\alias{CowBull.game}
\title{The Cow-Bull game}
\usage{
CowBull.game(num = NULL)
}
\arguments{
\item{num}{your number. If not specified computer will generate random number. If u r specifying your own number it is advisable to save it in a variable the press "cntrl+l" to clear console then call function and input that variable in the function. i.e if u save number x=1234 then press "cntrl+l" then give CowBull.game(x) and let your friend guess your number.}
}
\description{
The Cow-Bull game
}
\author{
Nithin Nayak
}
