% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronbach's.alpha.R
\name{cronbach.alpha}
\alias{cronbach.alpha}
\title{Cronbach's alpha}
\usage{
cronbach.alpha(data, f, n)
}
\arguments{
\item{data}{The Data frame}

\item{f}{The colum number of the first item}

\item{n}{The number of items}
}
\description{
Cronbach's alpha
}
\author{
Nithin Nayak
}
