% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.variables.R
\name{dendro.variables}
\alias{dendro.variables}
\title{foo}
\usage{
dendro.variables(data, method = c("associationMeasures", "distcor",
  "ClustOfVar"), linkage = "ward.D2", associationFun = association,
  check.psd = TRUE)
}
\description{
foo
}
