% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmap.R
\name{distmap}
\alias{distmap}
\title{foo}
\usage{
distmap(data, what = c("subjects", "variables"),
  variables.method = c("associationMeasures", "distcor"), varweights,
  linkage = "ward.D2", reorderdend, col, ...)
}
\description{
foo
}
