% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_extractor.R
\name{extract.data.frame}
\alias{extract.data.frame}
\title{Extracting data frame}
\usage{
extract.data.frame(data, sampleid, id)
}
\arguments{
\item{data}{the data frame which u want to extract the specific rows}

\item{sampleid}{The vector indicationg which values you need to extract of a particular column of the data}

\item{id}{The column number of the variable in the data frame which you want to compare with sample id to get extracted if the values are equal to the values in the sample id}
}
\description{
Extracting data frame
}
\author{
Nithin Nayak
}
