% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PasswordGenerator.R
\name{Gen.password}
\alias{Gen.password}
\title{Password generator}
\usage{
Gen.password(char = 8, spl.char = F, key = NULL, ppl = NULL)
}
\arguments{
\item{char}{number of character yo want in your password defaults to 8}

\item{spl.char}{set it to TRUE if u need special character in your password. Special characters are}

\item{key}{can be given any number which works as set.seed() if u need to create same password next time you run the same code}

\item{ppl}{number of password you want to create in one go (Works only if key is specified)}
}
\description{
Password generator
}
\author{
Nithin Nayak
}
