% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{install.all.packages}
\alias{install.all.packages}
\title{package installation}
\usage{
install.all.packages()
}
\arguments{
\item{.}{Nothing should be inputted}
}
\description{
package installation
}
\author{
Nithin Nayak
}
