% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.heatmap.R
\name{mix.heatmap}
\alias{mix.heatmap}
\title{mix heatmap}
\usage{
mix.heatmap(data, D.subjects, D.variables, dend.subjects, dend.variables,
  varweights, dist.variables.method = c("associationMeasures", "distcor",
  "ClustOfVar"), linkage = "ward.D2", associationFun = association,
  rowlab, rowmar = 3, lab.cex = 1.5, ColSideColors, RowSideColors,
  col.cont = marray::maPalette(low = "lightblue", high = "darkblue", k =
  50), cont.fixed.range = FALSE, cont.range, col.ord = list(low =
  "lightgreen", high = "darkgreen"), col.cat = c("indianred1", "darkred",
  "orangered", "orange", "palevioletred1", "violetred4", "red3",
  "indianred4"), legend.colbar, legend.rowbar, legend.mat = FALSE,
  legend.cex = 1, legend.srt = 0)
}
\description{
mix heatmap
}
