% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTMC (QUEUE PROBLEM).R
\name{MM1}
\alias{MM1}
\title{MM1 Queu Problem}
\usage{
MM1(lambda, mu, time)
}
\arguments{
\item{lambda}{Arrival rate}

\item{mu}{Departure rate}

\item{time}{time for which you want the process to run}
}
\description{
MM1 Queu Problem
}
\author{
Nithin
}
