% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTMC (QUEUE PROBLEM).R
\name{MM1Queue}
\alias{MM1Queue}
\title{MM1Queue}
\usage{
MM1Queue(N = 0, lambda, mu, Time = 0, n = 0)
}
\arguments{
\item{N}{initial size of the queue}

\item{lambda}{arrival rate}

\item{mu}{Departure rate}

\item{Time}{Time for which you want to see arrivals and departures in the queue}

\item{n}{number of transitions; either n or Time should be mentioned}
}
\description{
MM1Queue
}
\examples{
MM1Queue(lamda=3,mu=4,Time=8)
MM1Queue(lamda=4,mu=4,n=100)
}
\author{
Nithin Nayak
}
