% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RScriptsGenerator.R
\name{RscriptGenerator}
\alias{RscriptGenerator}
\title{R Script Files Generator Function}
\usage{
RscriptGenerator(path = getwd(), ScriptNames = NULL)
}
\arguments{
\item{path}{The directory path where you want to create the R script files. defaults to your current working directory.}

\item{ScriptNames}{The character vector indicating the names of the Rsripts to be generate.}
}
\description{
This function generates R script files in a given path.
}
\author{
Nithin Nayak
}
