% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.anova.test.R
\name{sample.anova.test}
\alias{sample.anova.test}
\title{Sample size for One-way ANOVA}
\usage{
sample.anova.test(k, d, sd, w = NULL, sig.level = 0.05, power = 0.8,
  nonpar = F, inflate = c(0.15, 0.2))
}
\arguments{
\item{k}{No. of groups (No.of multiple comparisons will be calculated automatically)}

\item{d}{The clinically significant difference}

\item{sd}{The estimate of standard deviation}

\item{w}{The effect size}

\item{sig.level}{The significance level defaults to 0.05}

\item{power}{Power of the test defaults to 0.8(80\%)}

\item{nonpar}{Should be set as nonpar=T if nonparametric inflation is needed.}

\item{inflate}{a vector or single value specifying the \% of inflation needed for nonparametric inflation default output is for 15\% and 20\%}
}
\description{
Gives the required minimum sample size for t test
}
\author{
Nithin Nayak
}
