% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.estimate.mean.R
\name{sample.estimate.mean}
\alias{sample.estimate.mean}
\title{Sample size for estimation of mean}
\usage{
sample.estimate.mean(d = NA, sd = NA, w = NA, sig.level = 0.05)
}
\arguments{
\item{d}{margin of error}

\item{sd}{Standard deviation}

\item{w}{effect size}

\item{sig.level}{Desierd significance level default is 0.05}
}
\value{
gives output n: minimum sample size
}
\description{
This command gives minimum sample size for estimation of mean and proportion
}
\author{
Nithin Nayak
}
