% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.estimate.prop.R
\name{sample.estimate.prop}
\alias{sample.estimate.prop}
\title{Sample size for estimation of proportion}
\usage{
sample.estimate.prop(p, d = NA, e = NA, sig.level = 0.05)
}
\arguments{
\item{p}{anticipated proportion}

\item{d}{margin or error}

\item{e}{relative precision}

\item{sig.level}{Desierd significance level default is 0.05}
}
\value{
gives output n: minimum sample size
}
\description{
This command gives minimum sample size for estimation of mean and proportion
}
\author{
Nithin Nayak
}
