% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.estimate.R
\name{sample.estimate}
\alias{sample.estimate}
\title{Sample size for estimation}
\usage{
sample.estimate(estimate, sig.level = 0.05)
}
\arguments{
\item{estimate}{The quantity you want to estimate either "mu" or "p"}

\item{sig.level}{Desierd significance level default is 0.05}
}
\value{
gives output n: minimum sample size
}
\description{
This command gives minimum sample size for estimation of mean and proportion
}
\author{
Nithin Nayak
}
