% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.kappa.estimate.R
\name{sample.kappa}
\alias{sample.kappa}
\title{Sample size for Kappa}
\usage{
sample.kappa(k, d, pi, sig.level = 0.05)
}
\arguments{
\item{k}{Anticipated kappa value}

\item{d}{margin of error}

\item{pi}{proportion of disagreement}
}
\value{
Returns the minimum sample size requierd for estimation for kappa coeffitient
}
\description{
Sample size for Kappa
}
\author{
Nithin Nayak
}
