% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.prop.test.R
\name{sample.prop.test}
\alias{sample.prop.test}
\title{Sample size for comparison of proportion}
\usage{
sample.prop.test(p1, p2, d = NA, sig.level = 0.05, power = 0.8,
  oneside = F)
}
\arguments{
\item{d}{The expected difference in proportion. If not entered d will be calculated as (p2-p1)}

\item{oneside}{Enter T if the test is onesided.}

\item{p1&p2}{estimate of proportion of first and second group respectively}

\item{sig.level&power}{Significance level and power of the test respectively.}
}
\value{
Returns the minimum sample size requierd per group
}
\description{
Sample size for comparison of proportion
}
\author{
Nithin Nayak
}
