% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.r.test.R
\name{sample.r.test}
\alias{sample.r.test}
\title{sample.r.test}
\usage{
sample.r.test(
  r = 0.3,
  sig.level = 0.05,
  power = 0.8,
  nonpar = F,
  inflate = c(0.15, 0.2)
)
}
\arguments{
\item{r}{Anticipated correlation coefficient. Defaults to 0.3}

\item{sig.level}{Significance level}

\item{power}{Requierd power of the test}

\item{nonpar}{Should be given as T if non parametric inflation is requierd}

\item{inflate}{Vector or single value indicting the \% of inflation}
}
\description{
sample.r.test
}
\author{
Nithin Nayak
}
