% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.repanova.test.R
\name{sample.repanova.test}
\alias{sample.repanova.test}
\title{Sample Size for repeated measures ANOVA}
\usage{
sample.repanova.test(d = NULL, sd = NULL, w = NULL, m, rho,
  sig.level = 0.05, power = 0.8, nonpar = F, inflate = c(0.15,
  0.2))
}
\arguments{
\item{d}{Clinically significant difference between groups}

\item{sd}{pooled standard deviation}

\item{w}{Effect size if estimates for sd or dis unknown}

\item{m}{Number of time points}

\item{rho}{Estimated correlation between outcomes measured at different time points}

\item{sig.level}{Significance level}

\item{power}{Power of the test}

\item{nonpar}{Set T if nonparametric inflation needed}

\item{inflate}{vector containing percentage of inflation needed default returns inflation for 15 and 20 \% inflation}
}
\description{
Sample Size for repeated measures ANOVA
}
\author{
Nithin Nayak
}
