% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.sen.spec.R
\name{sample.sen.spec}
\alias{sample.sen.spec}
\title{Sample size for Sensitivity Specificity}
\usage{
sample.sen.spec(sen = NULL, spec = NULL, d, p, sig.level = 0.05)
}
\arguments{
\item{sen}{The anticipated sensitivity}

\item{spec}{The anticipated specificity}

\item{d}{margin of error}

\item{p}{Prevelance}

\item{sig.level}{Level of Significance}
}
\description{
Sample size for Sensitivity Specificity
}
\author{
Nithin Nayak
}
