% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.t.test.R
\name{sample.t.test}
\alias{sample.t.test}
\title{Sample size for t test}
\usage{
sample.t.test(d = NA, sd = NA, w = NA, sig.level = 0.05,
  power = 0.8, oneside = F, nonpar = F, inflate = c(0.2, 0.15),
  paired = F)
}
\arguments{
\item{d}{The clinically significant difference}

\item{sd}{The estimate of standard deviation}

\item{w}{The effect size}

\item{sig.level}{The significance level defaults to 0.05}

\item{power}{Power of the test defaults to 0.8(80\%)}

\item{oneside}{Should set oneside=T if the test is onesided}

\item{nonpar}{Should be set as nonpar=T if nonparametric inflation is needed.}

\item{inflate}{a vector or single value specifying the \% of inflation needed for nonparametric inflation default output is for 15\% and 20\%}

\item{paired}{Should set paired=T if sample size is for paired t test}
}
\description{
Gives the required minimum sample size for t test
}
\author{
Nithin Nayak
}
