% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinySetup.R
\name{ShinySetup}
\alias{ShinySetup}
\title{File Structure generator function for Shiny App development}
\usage{
ShinySetup(path = getwd())
}
\arguments{
\item{path}{The directory path where you want to setup the Shiny app folder; defaults to your current working directory.}
}
\description{
This function help setup directories and files for new Shiny development.
}
\author{
Nithin Nayak
}
