% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.variables.R
\name{similarity.variables}
\alias{similarity.variables}
\title{foo}
\usage{
similarity.variables(data, method = c("associationMeasures", "distcor"),
  associationFun = association, check.psd = TRUE, make.psd = TRUE)
}
\description{
foo
}
