% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.help.R
\name{sample.help}
\alias{sample.help}
\title{List of codes available in the package}
\usage{
sample.help()
}
\description{
\link{sample.t.test}  For t-test (Paired,Independent,onesided)

\link{sample.prop.test} For difference in proportion

\link{sample.anova.test} For  One-Way anova

\link{sample.r.test} For correlation Coefficient

\link{sample.estimate} For estimation of mean or proportion
}
\author{
Nithin Nayak
}
